feature: C++20 modules with CMAKE, YES CMAKE!
