+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.20)
+cmake_minimum_required(VERSION 3.25)
 project(ray1 VERSION 1.0 LANGUAGES CXX)
 
 set(CMAKE_INCLUDE_CURRENT_DIR ON)
@@ -13,6 +13,11 @@ set(CMAKE_CXX_STANDARD 23)
 set(CMAKE_CXX_STANDARD_REQUIRED ON) 
 set(CMAKE_CXX_EXTENSIONS OFF)
 
+set(CMAKE_EXPERIMENTAL_CXX_MODULE_DYNDEP 1)
+# turn on the experimental API
+set(CMAKE_EXPERIMENTAL_CXX_MODULE_CMAKE_API  "3c375311-a3c9-4396-a187-3227ef642046")
+
+
 # Set up AUTOMOC and some sensible defaults for runtime execution
 # When using Qt 6.3, you can replace the code block below with
 # qt_standard_project_setup()
@@ -23,6 +28,7 @@ set(CMAKE_AUTOUIC ON)
 find_package(QT NAMES Qt5 REQUIRED COMPONENTS Core)
 find_package(Qt${QT_VERSION_MAJOR} REQUIRED COMPONENTS Concurrent Gui Widgets)
 
+
 add_executable(ray1 WIN32
     Ray.h
     calc.cpp calc.h
@@ -32,7 +38,13 @@ add_executable(ray1 WIN32
     rayview.cpp rayview.h rayview.ui
     fluid.cpp fluid.h fluid.ui
     shapes.h 
-    Module1.ixx)
+)
+
+target_sources(ray1 
+    PRIVATE
+        FILE_SET cxx_modules TYPE CXX_MODULES FILES
+        "FlipFluid.cxx"
+)
 
 target_link_libraries(ray1 PRIVATE
     Qt::Concurrent
