--- Module1.ixx
@@ -1,3 +1,0 @@
-export module Module1;
-
-export void MyFunc();
